new component- navbar
made a new navbar component. Cleaned app.js
