+++ client/src/components/Navbar/styles.js
@@ -1,0 +1,25 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export default makeStyles((theme) => ({
+    root: {
+        flexGrow: 1,
+      },
+      menuButton: {
+        marginRight: theme.spacing(2),
+      },
+      title: {
+        flexGrow: 1,
+        fontWeight: 'bold',
+      },
+      grow: {
+        flexGrow: 1,
+      },
+      
+      customizeAppbar: {
+        backgroundColor: 'white'
+    },
+    customMenuButton: {
+        color: '#2EA3F2',
+    },
+    
+    }));
